requirements.txt
# Pinned core project dependencies for reproducible setup with uv.
# Install after creating/activating a uv venv:
#   uv pip install -r requirements.txt

torch==2.10.0
numpy==2.2.6
pandas==2.3.3
polars==1.39.3
pyarrow==23.0.1
scikit-learn==1.8.0
PyYAML==6.0.3
tqdm==4.67.3
matplotlib==3.10.0
librosa==0.11.0
duckdb==1.5.1
transformers==4.57.6
huggingface-hub[cli]==0.36.2
safetensors==0.7.0
sentencepiece==0.2.1
